Add SSL support
This goes along with commits to Server::Starter and Starman. The first
to handle the ":ssl" suffix to the "IP:PORT::ssl" strings, and the
second to implement the maybe_upgrade_to_ssl method to handle the items
in $self->{options} where Starman::Server puts them.

Unit tests included.  I haven't found a way to quash this warning in the
test, but it doesn't appear when running IRL:

    Could not finalize SSL connection with client handle (SSL accept attempt
    failed error:0A000126:SSL routines::unexpected eof while reading)

I added a META.yml so at least the metacpan.org page would show a link
to the github repo.

Note that commit 5fccc0c27 "Merge pull request #2 from
nagata-logly/nagata-logly/support-for-unix-socket" was never released on
CPAN, so releasing this code would release that as well.